isad3000/VirtualJukeboxServer/requirements.txt

Django~=4.0.6
djangorestframework~=3.12.2
requests~=2.26.0
django-cors-headers
asgiref==3.5.2
psycopg2-binary==2.9.3
sqlparse==0.4.2
pytz
djangorestframework-gis
redis
geopy
channels==3.0.5
channels-redis==3.4.1
better-profanity
daphne
django-crontab
